+++ src/index.ts
@@ -1,1 +1,6 @@
 export { Radix, radix } from "./radix";
+export type {
+	Decodings,
+	Decoder,
+	RadixOptions
+} from "./types";