+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export { checkNumber } from "./check-number";
+export { decode } from "./decode";